Fix assets dir in devilutionX
The assets dir for devilutionX was changed some time ago to "assets_dvx" to avoid conflicts with other assets. The source of devilutionX is patched via a single sed-command in the package.mk file. Due to changes in the devilution source this sed-command patched the wrong line and in the end the assets were no longer found.