Amlogic-ne: fix cfgload
the current Generic_cfgload.src is kept the same as the one in CE,
however this means the fs label is not right for our bootup setup.

we're using `EmuELEC` as the fs label for the first partition on
disk, instead of `CoreELEC`. If kept the same the image won't boot,
changing that to `EmuELEC` instead.

should solve one of the problems mentioned in the following forum
post:

https://emuelec.discourse.group/t/trying-to-compile-for-s905w2-amlogic-ne/2013

quote @compusat22:
> first time it booted it fell over as it couldnt mount (coreelec partition)
> changed the AML_autoscript and cfgload from a prebuilt Amlogic-ng image. (as I think it was using ones from coreelec build)