Correction of supermariowar-data in the SuperMarioWar port:
Changes were made to the supermariowar-data source code, replacing the files that were previously .bmp now is .png.
I added a commit containing a version of supermariowar-data compatible with the EmuELEC version that uses .bmp files.
Additionally, I adjusted the "gptokeyb &" command by moving it outside the IF block. This is because, within the IF, the virtual keyboard would only start on the first execution. Now, outside the IF, it initializes correctly.